Bug 1080760 - Auto-generate bug suggestions asynchronously
This generates bug suggestions for ``text_log_summary`` artifacts submitted
to the ``artifact``  REST endpoint.  It schedules those asynchronously with
``celery``.

If a job is submitted to through the ``job`` REST endpoint and it contains a
``text_log_summary`` artifact with it, then the ``Bug suggestions`` artifact
is created immediately (not asynchronously).