Bug 1162682 - Catch exceptions whilst processing objectstore jobs
Now if an exception occurs during _load_ref_and_job_data_structs(), we
mark that job in the objectstore as errored and continue inserting the
other jobs. Previously the exception would have meant all of the other
jobs were not inserted, causing up to 100 rows in the objectstore to be
stuck in the 'loading' processed_state indefinitely.

The exception string passed to mark_object_error() isn't ideal, but it's
the same as the handling above, so will do for now until we remove the
objectstore.

In addition, this change means that we lose visibility in New Relic for
these exceptions - and someone has to manually check the objectstore for
jobs with error = "Y". However short term this seems preferable to
dropping 100 jobs every time we get an exception, particularly since
this is already the case for deserialisation exceptions. In a followup
bug we could always try using the New Relic Python agent's
record_exception() to maintain reporting without having to re-raise the
exception ourselves.